migration to aiogram 3.x

+++ modules/bot/handlers/default.py
@@ -1,40 +1,44 @@
-from aiogram import Dispatcher, types
-from aiogram.dispatcher import FSMContext
+from aiogram import Dispatcher, F, types
+from aiogram.filters.command import Command
 
-from ...logger import logger, print_msg
-from ..keyboards.default import add_delete_button
+from modules.bot.keyboards.default import add_delete_button
+from modules.logger import Logger
 
 
-@print_msg
-async def start(message: types.Message, state: FSMContext):
+@Logger.log_msg
+async def start(message: types.Message) -> None:
     text = "Start message"
-    await message.reply(text, reply_markup=add_delete_button())
-    await state.finish()
+    await message.reply(text, reply_markup=add_delete_button().as_markup())
+    return None
 
 
-@print_msg
-async def help(message: types.Message):
+@Logger.log_msg
+async def help_handler(message: types.Message) -> None:
     text = "Help message"
-    await message.reply(text, reply_markup=add_delete_button())
+    await message.reply(text, reply_markup=add_delete_button().as_markup())
+    return
 
 
-async def delete_msg(query: types.CallbackQuery):
+async def delete_msg(query: types.CallbackQuery) -> None:
+    if not isinstance(query.message, types.Message):
+        return
+    if query.bot is None:
+        return
     try:
         await query.bot.delete_message(query.message.chat.id, query.message.message_id)
         if query.message.reply_to_message:
             await query.bot.delete_message(query.message.chat.id, query.message.reply_to_message.message_id)
         await query.answer()
     except Exception as exc:
-        logger.error(exc)
+        Logger.error(exc)
         await query.answer("Error")
 
 
-def register_handlers_default(dp: Dispatcher):
-    dp.register_message_handler(start, commands="start", state="*")
-    dp.register_message_handler(help, commands="help", state="*")
+def register_handlers_default(dp: Dispatcher) -> None:
+    dp.message.register(start, Command("start"))
+    dp.message.register(help_handler, Command("help"))
 
-    dp.register_callback_query_handler(
+    dp.callback_query.register(
         delete_msg,
-        lambda c: c.data == "delete",
-        state="*"
+        F.func(lambda c: c.data == "delete"),
     )
